% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvch_marginal.R
\name{Marginals}
\alias{Marginals}
\alias{dmarg1}
\alias{dmarg2}
\alias{pmarg1}
\alias{pmarg2}
\title{Marginal Distributions of the Bivariate Chen Law}
\usage{
dmarg1(x, alpha1, alpha2, alpha3, beta)

dmarg2(y, alpha1, alpha2, alpha3, beta)

pmarg1(q, alpha1, alpha2, alpha3, beta)

pmarg2(q, alpha1, alpha2, alpha3, beta)
}
\arguments{
\item{x, q}{non-negative quantiles.}

\item{alpha1, alpha2, alpha3}{positive shape parameters of the three
latent Chen variables.}

\item{beta}{positive common power parameter.}

\item{y}{quantile at which the second marginal is evaluated.}
}
\value{
Numeric vectors of density (\code{d*}) or probability
  (\code{p*}) values.
}
\description{
Density and CDF of the marginals
  \eqn{Z_1 \sim Ch(\alpha_1+\alpha_3,\beta)} and
  \eqn{Z_2 \sim Ch(\alpha_2+\alpha_3,\beta)}.
}
\examples{
dmarg1(1, 1, 1, 1, 1.5)
dchen(1, 2, 1.5)                   # identical by construction
pmarg2(c(0.5, 1, 2), 1, 1, 1, 1.5)
pchen(c(0.5, 1, 2), 2, 1.5)

}
