% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvch_plots.R
\name{plot.bvch}
\alias{plot.bvch}
\title{Plot Method for Bivariate Chen Objects}
\usage{
\method{plot}{bvch}(x, alpha1 = NULL, alpha2 = NULL, alpha3 = NULL, beta = NULL, ...)
}
\arguments{
\item{x}{object of class \code{"bvch"} or \code{"bvchfit"}.}

\item{alpha1, alpha2, alpha3, beta}{distribution parameters (required
when \code{x} is a data matrix).}

\item{...}{further arguments passed to methods.}
}
\description{
Draws a contour of the absolutely continuous joint density with the
simulated points (or the fitted parameters) superimposed.
}
