% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BvChen-package.R
\docType{package}
\name{BvChen-package}
\alias{BvChen}
\alias{BvChen-package}
\title{BvChen: The Bivariate Chen Distribution}
\description{
Density, distribution function, survival function, random variate generation, marginal and conditional distributions, dependence measures, moments, maximum-likelihood estimation via the EM algorithm and goodness-of-fit tools for the bivariate Chen distribution constructed as min(Z1, Z2) of three independent Chen (2000) components. Includes univariate Chen building blocks.
}
\details{
Help and issue reporting are available at the URLs below.
\itemize{
  \item \url{https://github.com/MukulBijalwan/BvChen}
  \item Report bugs at \url{https://github.com/MukulBijalwan/BvChen/issues}
}
}
\keyword{internal}
